Check individual messages for distinct sources
Instead of using the PostCSS Result's source and grouping all
messages under that, check each individual message for a node
with a distinct source, group messages according to whether
their nodes' share sources, and only use the Result's source
if the message does not have its own node with a source.

This addresses #12.